requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
langchain==0.1.11
langchain-google-genai==1.0.1
langchain-community==0.0.27
langchain-chroma==0.1.0
pypdf==4.1.0
python-docx==1.1.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
chromadb==0.4.24
tiktoken==0.6.0
langchain-experimental==0.0.54
loguru==0.7.2
python-multipart==0.0.9
